clean-up more task code
