+++ api_tests/nodes/views/test_node_detail_update.py
@@ -92,8 +92,7 @@ class TestNodeUpdate(NodeCRUDTestCase):
         project_private.add_contributor(
             non_admin,
             permissions=permissions.WRITE,
-            auth=Auth(project_private.creator),
-            notification_type=False
+            auth=Auth(project_private.creator)
         )
         project_private.save()
         res = app.patch_json(
@@ -113,9 +112,7 @@ class TestNodeUpdate(NodeCRUDTestCase):
             project_private.add_contributor(
                 admin_user,
                 permissions=permissions.ADMIN,
-                auth=Auth(project_private.creator),
-                notification_type=False
-            )
+                auth=Auth(project_private.creator))
             project_private.save()
             with capture_notifications():
                 res = app.patch_json_api(
